Fix according code review * format imports * move network code from deprecated entity * remove not needed resource attribute * fixed setting values to ft state * removed not needed comment
