+++ vcd/resource_vcd_network_routed.go
@@ -1,7 +1,12 @@
 package vcd
 
 import (
+	"bytes"
 	"fmt"
+	"log"
+	"strings"
+
+	"github.com/hashicorp/terraform/helper/hashcode"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/vmware/go-vcloud-director/govcd"
@@ -22,13 +27,11 @@ func resourceVcdNetworkRouted() *schema.Resource {
 			},
 			"org": {
 				Type:     schema.TypeString,
-				Required: false,
 				Optional: true,
 				ForceNew: true,
 			},
 			"vdc": {
 				Type:     schema.TypeString,
-				Required: false,
 				Optional: true,
 				ForceNew: true,
 			},
@@ -48,7 +51,6 @@ func resourceVcdNetworkRouted() *schema.Resource {
 
 			"gateway": &schema.Schema{
 				Type:     schema.TypeString,
-				Required: false,
 				Optional: true,
 				ForceNew: true,
 			},
@@ -229,4 +231,75 @@ func resourceVcdNetworkRoutedCreate(d *schema.ResourceData, meta interface{}) er
 	d.SetId(d.Get("name").(string))
 
 	return resourceVcdNetworkRead(d, meta)
+}
+
+func resourceVcdNetworkRead(d *schema.ResourceData, meta interface{}) error {
+	vcdClient := meta.(*VCDClient)
+
+	_, vdc, err := vcdClient.GetOrgAndVdcFromResource(d)
+	if err != nil {
+		return fmt.Errorf(errorRetrievingOrgAndVdc, err)
+	}
+
+	network, err := vdc.FindVDCNetwork(d.Id())
+	if err != nil {
+		log.Printf("[DEBUG] Network no longer exists. Removing from tfstate")
+		d.SetId("")
+		return nil
+	}
+
+	d.Set("name", network.OrgVDCNetwork.Name)
+	d.Set("href", network.OrgVDCNetwork.HREF)
+	if c := network.OrgVDCNetwork.Configuration; c != nil {
+		d.Set("fence_mode", c.FenceMode)
+		if c.IPScopes != nil {
+			d.Set("gateway", c.IPScopes.IPScope.Gateway)
+			d.Set("netmask", c.IPScopes.IPScope.Netmask)
+			d.Set("dns1", c.IPScopes.IPScope.DNS1)
+			d.Set("dns2", c.IPScopes.IPScope.DNS2)
+		}
+	}
+
+	return nil
+}
+
+func resourceVcdNetworkDelete(d *schema.ResourceData, meta interface{}) error {
+	vcdClient := meta.(*VCDClient)
+	vcdClient.Mutex.Lock()
+	defer vcdClient.Mutex.Unlock()
+
+	_, vdc, err := vcdClient.GetOrgAndVdcFromResource(d)
+	if err != nil {
+		return fmt.Errorf(errorRetrievingOrgAndVdc, err)
+	}
+
+	network, err := vdc.FindVDCNetwork(d.Id())
+	if err != nil {
+		return fmt.Errorf("error finding network: %#v", err)
+	}
+
+	err = retryCall(vcdClient.MaxRetryTimeout, func() *resource.RetryError {
+		task, err := network.Delete()
+		if err != nil {
+			return resource.RetryableError(
+				fmt.Errorf("error Deleting Network: %#v", err))
+		}
+		return resource.RetryableError(task.WaitTaskCompletion())
+	})
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func resourceVcdNetworkIPAddressHash(v interface{}) int {
+	var buf bytes.Buffer
+	m := v.(map[string]interface{})
+	buf.WriteString(fmt.Sprintf("%s-",
+		strings.ToLower(m["start_address"].(string))))
+	buf.WriteString(fmt.Sprintf("%s-",
+		strings.ToLower(m["end_address"].(string))))
+
+	return hashcode.String(buf.String())
 }
